Updates based on PR feedback
