+++ base/collection_view.go
@@ -204,81 +204,6 @@ func (c *Collection) executeViewQuery(ddoc, name string, params map[string]inter
 	return goCbViewResult.Raw(), nil
 }
 
-type gocbResultRaw interface {
-
-	// NextBytes returns the next row as bytes.
-	NextBytes() []byte
-
-	// Err returns any errors that have occurred on the stream
-	Err() error
-
-	// Close marks the results as closed, returning any errors that occurred during reading the results.
-	Close() error
-
-	// MetaData returns any meta-data that was available from this query as bytes.
-	MetaData() ([]byte, error)
-}
-
-// GoCBQueryIterator wraps a gocb v2 ViewResultRaw to implement sgbucket.QueryResultIterator
-type gocbRawIterator struct {
-	rawResult gocbResultRaw
-}
-
-func NewGoCBQueryIterator(viewResult *gocb.ViewResultRaw) *gocbRawIterator {
-	return &gocbRawIterator{
-		rawResult: viewResult,
-	}
-}
-
-// Unmarshal a single result row into valuePtr, and then close the iterator
-func (i *gocbRawIterator) One(valuePtr interface{}) error {
-	if !i.Next(valuePtr) {
-		err := i.Close()
-		if err != nil {
-			return nil
-		}
-		return gocb.ErrNoResult
-	}
-
-	// Ignore any errors occurring after we already have our result
-	//  - follows approach used by gocb v1 One() implementation
-	_ = i.Close()
-	return nil
-}
-
-// Unmarshal the next result row into valuePtr.  Returns false when reaching end of result set
-func (i *gocbRawIterator) Next(valuePtr interface{}) bool {
-
-	nextBytes := i.rawResult.NextBytes()
-	if nextBytes == nil {
-		return false
-	}
-
-	err := JSONUnmarshal(nextBytes, &valuePtr)
-	if err != nil {
-		Warnf("Unable to marshal view result row into value: %v", err)
-		return false
-	}
-	return true
-}
-
-// Retrieve raw bytes for the next result row
-func (i *gocbRawIterator) NextBytes() []byte {
-	return i.rawResult.NextBytes()
-}
-
-// Closes the iterator.  Returns any row-level errors seen during iteration.
-func (i *gocbRawIterator) Close() error {
-
-	// check for errors before closing?
-	closeErr := i.rawResult.Close()
-	if closeErr != nil {
-		return closeErr
-	}
-	resultErr := i.rawResult.Err()
-	return resultErr
-}
-
 // waitForAvailableViewOp prevents Sync Gateway from having too many concurrent
 // view queries against Couchbase Server
 func (c *Collection) waitForAvailViewOp() {
